src/pyproject.toml
[project]
name = "brainrushbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aerich>=0.8.2",
    "aiogram>=3.20.0",
    "asyncpg>=0.30.0",
    "pydantic-settings>=2.8.1",
    "tortoise-orm>=0.25.0",
    "watchfiles>=1.0.5",
]

[tool.aerich]
tortoise_orm = "config.AERICH_CONFIG"
location = "./migrations"
src_folder = "./"
